owasp-top10-2017-apps/a8/amarelo-designs/app/requirements.txt
flask
Flask
PyJWT==1.7.1